Merged/21 05 2024

+++ src/main/java/school/sptech/backend/domain/cesta/repository/CestaRepository.java
@@ -3,5 +3,5 @@ package school.sptech.backend.domain.cesta.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import school.sptech.backend.domain.cesta.Cesta;
 
-public interface CestaRepository extends JpaRepository<Cesta, Long> {
+public interface CestaRepository extends JpaRepository<Cesta, Integer> {
 }
